pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["askhat <askhat.9812@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.2"
pydantic-settings = "^2.0.3"
email-validator = "^2.0.0.post2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-jwt = "^0.1.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "database.DATABASE_CONFIG"
location = "migrations"
src_folder = "./src"
